test max clique

+++ src/MemoryManager.h
@@ -104,7 +104,7 @@ static void* CallocWithCheck(size_t x, size_t y)
     #else
 
     #define Malloc(x) malloc(max(x,1))
-    #define Calloc(x,y) calloc(max(x,1),max(y,1))
+    #define Calloc(x,y) calloc(max((int)x,1),max((int)y,1))
     #define Free(x) free(x)
 
     #endif // ALLOW_ALLOC_ZERO_BYTES
